Next:
* lang valid values start now from 0, not from -1. -1 was useless,
  as the only place it had some influence was buffer staff. Still,
  even there -1 forced skipping of GETMSG call, which in case of
  lang == 0 would have returned the same string (mlstr_mval). To
  avoid searching msgdb in this case, check for lang == 0 early in
  GETMSG() and substitute buf_new(-1) calls with buf_new(0)

* changed lang_t.slang_of semantics, in case of "no slang" it
  references itself. By the way, it _was_ broken in case of
  deleting not last lang, if there were slang_of pointers to
  languages after it, e.g. langs 0, 1, 2, 3, with 3 having
  slang_of == 2 and deleting 1. After deletion, 3 would became 2.
  Luckily, we have no lang_delete. :) But in case we will have,
  implement lang_destroy.

* MLSTR_FOREACH_FUN(name) declaration to declare mlstr_foreach callbacks

* Coredump in mod_olc fixed (exd format), it passed NULL callback to
  mlstr_foreach. Also, cb_format did not return a value, I suspect it
  could cause stack corruption when mlstr_forech tries to access its
  return value.

ABI_VERSION bump (changed buf_new ABI)


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@2868 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/include/buffer.h
@@ -1,5 +1,5 @@
 /*
- * $Id: buffer.h,v 1.15 2001-06-16 18:40:08 fjoe Exp $
+ * $Id: buffer.h,v 1.16 2001-06-22 07:13:32 avn Exp $
  */
 
 /***************************************************************************
@@ -46,7 +46,7 @@
 #define BUF_START	1
 #define BUF_END		0
 
-BUFFER *buf_new		(int lang);
+BUFFER *buf_new		(size_t lang);
 void	buf_free	(BUFFER *buffer);
 
 bool	buf_prepend	(BUFFER *buffer, const char *string);
@@ -66,7 +66,7 @@ bool	buf_act3	(BUFFER *buffer, int where, const char *format,
 
 void	buf_clear	(BUFFER *buffer);
 char *	buf_string	(BUFFER *buffer);
-int	buf_lang	(BUFFER *buffer);
+size_t	buf_lang	(BUFFER *buffer);
 
 extern int nAllocBuf;
 extern int sAllocBuf;
